Rename plugin to Daily Tree Runs
