+++ build.gradle
@@ -26,7 +26,7 @@ dependencies {
 	testImplementation group: 'net.runelite', name:'jshell', version: runeLiteVersion
 }
 
-group = 'com.example'
+group = 'com.dailytree'
 version = '1.0-SNAPSHOT'
 
 tasks.withType(JavaCompile).configureEach {
@@ -36,9 +36,9 @@ tasks.withType(JavaCompile).configureEach {
 
 tasks.register('shadowJar', Jar) {
 	dependsOn configurations.testRuntimeClasspath
-	manifest {
-		attributes('Main-Class': 'com.example.ExamplePluginTest', 'Multi-Release': true)
-	}
+        manifest {
+                attributes('Main-Class': 'com.dailytree.DailyTreeRunsPluginTest', 'Multi-Release': true)
+        }
 
 	duplicatesStrategy = DuplicatesStrategy.EXCLUDE
 	from sourceSets.main.output
