+++ runelite-plugin.properties
@@ -1,5 +1,5 @@
-displayName=Example
-author=Nobody
-description=An example greeter plugin
+displayName=Daily Tree Runs
+author=Farmboy
+description=Track your daily tree runs
 tags=
-plugins=com.example.ExamplePlugin
+plugins=com.dailytree.DailyTreeRunsPlugin
