+++ src/main/java/com/dailytree/DailyTreeRunsConfig.java
@@ -1,11 +1,11 @@
-package com.example;
+package com.dailytree;
 
 import net.runelite.client.config.Config;
 import net.runelite.client.config.ConfigGroup;
 import net.runelite.client.config.ConfigItem;
 
-@ConfigGroup("example")
-public interface ExampleConfig extends Config
+@ConfigGroup("dailytreeruns")
+public interface DailyTreeRunsConfig extends Config
 {
 	@ConfigItem(
 		keyName = "greeting",
