+++ src/main/java/com/dailytree/DailyTreeRunsPlugin.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.dailytree;
 
 import com.google.inject.Provides;
 import javax.inject.Inject;
@@ -14,26 +14,26 @@ import net.runelite.client.plugins.PluginDescriptor;
 
 @Slf4j
 @PluginDescriptor(
-	name = "Example"
+        name = "Daily Tree Runs"
 )
-public class ExamplePlugin extends Plugin
+public class DailyTreeRunsPlugin extends Plugin
 {
 	@Inject
 	private Client client;
 
 	@Inject
-	private ExampleConfig config;
+        private DailyTreeRunsConfig config;
 
 	@Override
 	protected void startUp() throws Exception
 	{
-		log.info("Example started!");
+                log.info("Daily Tree Runs started!");
 	}
 
 	@Override
 	protected void shutDown() throws Exception
 	{
-		log.info("Example stopped!");
+                log.info("Daily Tree Runs stopped!");
 	}
 
 	@Subscribe
@@ -41,13 +41,13 @@ public class ExamplePlugin extends Plugin
 	{
 		if (gameStateChanged.getGameState() == GameState.LOGGED_IN)
 		{
-			client.addChatMessage(ChatMessageType.GAMEMESSAGE, "", "Example says " + config.greeting(), null);
+                        client.addChatMessage(ChatMessageType.GAMEMESSAGE, "", "Daily Tree Runs says " + config.greeting(), null);
 		}
 	}
 
 	@Provides
-	ExampleConfig provideConfig(ConfigManager configManager)
-	{
-		return configManager.getConfig(ExampleConfig.class);
-	}
+        DailyTreeRunsConfig provideConfig(ConfigManager configManager)
+        {
+                return configManager.getConfig(DailyTreeRunsConfig.class);
+        }
 }
